libspl: - Експерементальные каналы(пайпы) - Убрана задержка в цсм при создании. Задержка при необходимости будет выполнятся на стадии закрытия(в случае убийства цсм сразу после создания) - В таймеры добавлен параметр userdata - Добавление процессовых и тредовых списков производится под контролем мютекса - Убрана(пока) поддержка исключений. Из за них иногда происходили непонятные пики - В управление памяти переделан принцып базы указателей, чуть медленней предидущего, но более стабильный к порче памяти - Добавлен контроль AllocWS/FreeWS - Удалены хедеры из папки проекта, все перенесено в инклуды - В свилибе расширяется режим поддержки spl(__USE_SPL__), при котором некоторые вызовы подменяются на библиотечные вызовы
RadioSi:
- Добавлена поддержка фонового буфферизатор через пайпы. Отдельный тред качает поток паралельно с программой(декодером етц)
- Вывод текущего вида деятельности(дкодирование, буфферизирование)
- Протестил на двух других станциях, декодер faad2 не осилил декодировать поток о_О(стандартный плеер тоже)

+++ tests/RadioSI/faad/src/decoder.c
@@ -1,19 +1,19 @@
 /*
 ** FAAD2 - Freeware Advanced Audio (AAC) Decoder including SBR decoding
 ** Copyright (C) 2003-2005 M. Bakker, Nero AG, http://www.nero.com
-**  
+**
 ** This program is free software; you can redistribute it and/or modify
 ** it under the terms of the GNU General Public License as published by
 ** the Free Software Foundation; either version 2 of the License, or
 ** (at your option) any later version.
-** 
+**
 ** This program is distributed in the hope that it will be useful,
 ** but WITHOUT ANY WARRANTY; without even the implied warranty of
 ** MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 ** GNU General Public License for more details.
-** 
+**
 ** You should have received a copy of the GNU General Public License
-** along with this program; if not, write to the Free Software 
+** along with this program; if not, write to the Free Software
 ** Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
 **
 ** Any non-GPL usage of this software or parts of this software is strictly
@@ -264,7 +264,7 @@ long NEAACDECAPI NeAACDecInit(NeAACDecHandle hpDecoder,
 #endif
 
         faad_initbits(&ld, buffer, buffer_size);
- 
+
 #if 0
         memset(l, 0, sizeof(latm_header));
         is_latm = latmCheck(l, &ld);
@@ -1094,7 +1094,7 @@ static void* aac_frame_decode(NeAACDecStruct *hDecoder,
                 faad_free(hDecoder->sample_buffer);
             hDecoder->sample_buffer = NULL;
             hDecoder->sample_buffer = faad_malloc(frame_len*output_channels*stride);
-        } else if (sample_buffer_size < frame_len*output_channels*stride) {
+        } else if ((long)sample_buffer_size < frame_len*output_channels*stride) {
             /* provided sample buffer is not big enough */
             hInfo->error = 27;
             return NULL;
